Move and rename the file to RTClient
